pyproject.toml
[project]
name = "discord-playlistbot"
version = "0.1.0"
description = "Discord音楽リンク収集Bot - YouTubeとSoundCloudのリンクを自動でプレイリストに追加"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "discord.py>=2.3.0",
    "google-api-python-client>=2.100.0",
    "google-auth-oauthlib>=1.0.0",
    "aiohttp>=3.8.0",
    "aiosqlite>=0.19.0",
    "python-dotenv>=1.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["ALL"]
ignore = []

[tool.ruff.format]
quote-style = "double"
